requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
virtualenv==15.0.3
djangorestframework==3.4.6
django-crispy-forms==1.6.0
django-filter==0.14.0